+++ commands.go
@@ -1,3 +1,6 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package raft
 
 // RPCHeader is a common sub-structure used to pass along protocol version and