+++ file_snapshot_test.go
@@ -1,9 +1,11 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
 package raft
 
 import (
 	"bytes"
 	"io"
-	"io/ioutil"
 	"os"
 	"reflect"
 	"runtime"
@@ -25,13 +27,13 @@ func TestFileSnapshotSinkImpl(t *testing.T) {
 }
 
 func TestFileSS_CreateSnapshotMissingParentDir(t *testing.T) {
-	parent, err := ioutil.TempDir("", "raft")
+	parent, err := os.MkdirTemp("", "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}
 	defer os.RemoveAll(parent)
 
-	dir, err := ioutil.TempDir(parent, "raft")
+	dir, err := os.MkdirTemp(parent, "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}
@@ -47,11 +49,11 @@ func TestFileSS_CreateSnapshotMissingParentDir(t *testing.T) {
 	if err != nil {
 		t.Fatalf("should not fail when using non existing parent")
 	}
-
 }
+
 func TestFileSS_CreateSnapshot(t *testing.T) {
 	// Create a test dir
-	dir, err := ioutil.TempDir("", "raft")
+	dir, err := os.MkdirTemp("", "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}
@@ -159,7 +161,7 @@ func TestFileSS_CreateSnapshot(t *testing.T) {
 
 func TestFileSS_CancelSnapshot(t *testing.T) {
 	// Create a test dir
-	dir, err := ioutil.TempDir("", "raft")
+	dir, err := os.MkdirTemp("", "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}
@@ -197,7 +199,7 @@ func TestFileSS_Retention(t *testing.T) {
 	var err error
 	// Create a test dir
 	var dir string
-	dir, err = ioutil.TempDir("", "raft")
+	dir, err = os.MkdirTemp("", "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}
@@ -250,7 +252,7 @@ func TestFileSS_BadPerm(t *testing.T) {
 
 	// Create a temp dir
 	var dir1 string
-	dir1, err = ioutil.TempDir("", "raft")
+	dir1, err = os.MkdirTemp("", "raft")
 	if err != nil {
 		t.Fatalf("err: %s", err)
 	}
@@ -258,11 +260,11 @@ func TestFileSS_BadPerm(t *testing.T) {
 
 	// Create a sub dir and remove all permissions
 	var dir2 string
-	dir2, err = ioutil.TempDir(dir1, "badperm")
+	dir2, err = os.MkdirTemp(dir1, "badperm")
 	if err != nil {
 		t.Fatalf("err: %s", err)
 	}
-	if err = os.Chmod(dir2, 000); err != nil {
+	if err = os.Chmod(dir2, 0o00); err != nil {
 		t.Fatalf("err: %s", err)
 	}
 	defer os.Chmod(dir2, 777) // Set perms back for delete
@@ -274,13 +276,13 @@ func TestFileSS_BadPerm(t *testing.T) {
 }
 
 func TestFileSS_MissingParentDir(t *testing.T) {
-	parent, err := ioutil.TempDir("", "raft")
+	parent, err := os.MkdirTemp("", "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}
 	defer os.RemoveAll(parent)
 
-	dir, err := ioutil.TempDir(parent, "raft")
+	dir, err := os.MkdirTemp(parent, "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}
@@ -294,7 +296,7 @@ func TestFileSS_MissingParentDir(t *testing.T) {
 
 func TestFileSS_Ordering(t *testing.T) {
 	// Create a test dir
-	dir, err := ioutil.TempDir("", "raft")
+	dir, err := os.MkdirTemp("", "raft")
 	if err != nil {
 		t.Fatalf("err: %v ", err)
 	}